optimised ProjectHero
